Реализовано полное тестовое покрытие IEntity и Entity Исправлено поведение Handle при подаче невалидного ID Добавлена защита от использования удалённых дочерних объектов в IEntity Добавлен метод ReserveChildSpace в IEntity для резервирования памяти под потомков Добавлены методы GetParent и GetChild в IEntity для получения объекта Добавлена логика в IComponent связанная с изменением дерева Добавлен метод UpdateTreeComponents для оповещения о изменении родителя сущности её компонентов Реализован компонент Transform

+++ include/ECS/util/Timer.h
@@ -31,7 +31,7 @@ namespace ECS::util {
             return elapsed.count();
         }
 
-        void SetTimeScale(f64 ts) noexcept;
+        void SetTimeScale(f32 ts) noexcept;
 
         [[nodiscard]] inline f32 GetTimeScale() const noexcept
         {
